\name{E2.1}
\alias{E2.1}
\non_function{}
\title{ Data on Grade Point Average and SAT Scores }
\usage{data(E2.1)}
\description{
The \code{E2.1} data frame has 9 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{GPA}{
a numeric vector, grade point average (maximum=4)
}
\item{SATV}{
a numeric vector, SAT verbal score.
}
\item{SATM}{
a numeric vector, SAT mathematical score.
}
}
}
\source{
 Dacey (1983).
}
\examples{
data(E2.1)
}
\keyword{datasets}
