\name{E2.8}
\alias{E2.8}
\title{ Data on per Capita Output of Workers in Shanghai }
\concept{Data on per Capita Output of Workers in Shanghai}
\usage{data(E2.8)}
\description{
The \code{E2.8} data frame has 17 rows and 4 columns, data
 for 17 factories in Shanghai.
}
\format{
This data frame contains the following columns:
\describe{
\item{Output}{
a numeric vector, per capita output in Chinese Yuan.
}
\item{SI}{
a numeric vector, number of workers in the factory.
}
\item{SP}{
a numeric vector, land area of the factory in sq. meters per worker.
}
\item{I}{
a numeric vector, investments in Yuan per worker.
}
}
}
\source{
 Prof. Zhang Tingwei of Tongji University, Shanghai.
}
\examples{
   data(E2.8)
   summary(E2.8)
}
\keyword{datasets}
\concept{regression}


