\name{E3.4}
\alias{E3.4}
\title{ Men's Worlds Record Times for Running and Corresponding Distances }
\concept{ Men's Worlds Record Times for Running and Corresponding Distances }
\usage{data(E3.4)}
\description{
The \code{E3.4} data frame has 13 rows and 2 columns.
 World record times as of 1974.
}
\format{
This data frame contains the following columns:
\describe{
\item{Dist.}{
a numeric vector, distance in meters.
}
\item{Time}{
a numeric vector, time in seconds.
}
}
}
\seealso{ \code{\link{E3.5}}, the records for women. }
\source{
 Encyclopædia Britannica, 15th Edition, 1974, Micropædia, IX, 
 page 485. 
}
\examples{
data(E3.4)
summary(E3.4)
}
\keyword{datasets}
\concept{regression}

