\name{E3.6}
\alias{E3.6}
\non_function{}
\title{ Data on Corporations and Corporation Chairmen }
\concept{ Data on Corporations and Corporation Chairmen }
\usage{data(E3.6)}
\description{
The \code{E3.6} data frame has 50 rows and 6 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Y84}{
a numeric vector, salary 1984, in dollars.
}
\item{Y83}{
a numeric vector, salary 1983, in dollars.
}
\item{SHARES}{
a numeric vector, number of shares the chairman holds.
}
\item{REV}{
a numeric vector, total revenue of the company.
}
\item{INC}{
a numeric vector, total income of the company.
}
\item{AGE}{
a numeric vector, age of chairman, in years.
}
}
}
\source{
 Reprinted with permission from the May 13, 1985, issue of  Crain's
 Chicago Business. Copyright 1985 by Crain's Communications, Inc. The
 data given are a portion of the original table.  
}
\examples{
data(E3.6)
summary(E3.6)
}
\keyword{datasets}
\concept{regression}

