\name{E3.9}
\alias{E3.9}
\non_function{}
\title{ Blood Velocity Data }
\usage{data(E3.9)}
\description{
The \code{E3.9} data frame has 18 rows and 4 columns.
All the observations are for the same person.
}
\format{
This data frame contains the following columns:
\describe{
\item{x.1}{
a numeric vector, cardiac output.
}
\item{x.2}{
a numeric vector, carbon dioxide level in the blood.
}
\item{y}{
a numeric vector, blood flow velocity in the brain.
}
\item{Aminophylline}{
a factor with levels
\code{no} 
\code{with}, Aminophylline used or not. The hypothesis is that 
 aminophylline retards blood flow. 
}
}
}
\source{
 Tonse Raju, M.D., Department of Neonatology, University of Illinois
 at Chicago.
}
\examples{
data(E3.9)
}
\keyword{datasets}
