\name{E6.11}
\alias{E6.11}
\non_function{}
\title{ Heights of Fathers and Sons }
\usage{data(E6.11)}
\description{
The \code{E6.11} data frame has 12 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Height.Father}{
a numeric vector, height of father to the nearest inch.
}
\item{Aver.Height.Son}{
a numeric vector, average heights of sons.
}
\item{No.Fathers}{
a numeric vector, number of fathers in each group.
}
}
}

\source{
 dacey (1983, Ch. 1) from McNemar (1969, p. 130), \emph{Psycological Statistics}.
}
\examples{
data(E6.11)
}
\keyword{datasets}
