\name{E7.1}
\alias{E7.1}
\non_function{}
\title{ Data on Dental Measurements }
\concept{ Data on Dental Measurements }
\concept{growth curve}
\usage{data(E7.1)}
\description{
The \code{E7.1} data frame has 4 rows and 12 columns.
Dental measurements for girls from 8 to 14 years old. Each measurement is the
distance, in mm, from the center of the pituary to the ptery-maxilliary fissure.
}
\format{
This data frame contains the following columns:
\describe{
\item{Age}{
a numeric vector, age of girl when measurement was taken.
}
\item{S.1}{
a numeric vector, measurements for girl 1.
}
\item{S.2}{
a numeric vector, measurements for girl 2.
}
\item{S.3}{
a numeric vector, measurements for girl 3.
}
\item{S.4}{
a numeric vector, measurements for girl 4.
}
\item{S.5}{
a numeric vector, measurements for girl 5.
}
\item{S.6}{
a numeric vector, measurements for girl 6.
}
\item{S.7}{
a numeric vector, measurements for girl 7.
}
\item{S.8}{
a numeric vector, measurements for girl 8.
}
\item{S.9}{
a numeric vector, measurements for girl 9.
}
\item{S.10}{
a numeric vector, measurements for girl 10.
}
\item{S.11}{
a numeric vector, measurements for girl 11.
}
}
}
\source{
 Pothoff and Roy (1964).
}
\examples{
data(E7.1)
summary(E7.1)
}
\keyword{datasets}
