\name{E7.2}
\alias{E7.2}
\non_function{}
\title{ Prices of Crude Oil, Natural Gas, Bituminous Coal and Lignite, 
           and Anthracite by Year. }
\concept{Prices of Crude Oil, Natural Gas, Bituminous Coal and Lignite, and Anthracite by Year.}
\usage{data(E7.2)}
\description{
The \code{E7.2} data frame has 32 rows and 5 columns.
 Prices are in 1972 cents (U.S) by 1000 BTU.}
\format{
This data frame contains the following columns:
\describe{
\item{year}{
a numeric vector, year of observation.
}
\item{Oil}{
a numeric vector, price of oil.
}
\item{Gas}{
a numeric vector, price of Gas.
}
\item{Bit.}{
a numeric vector, price of Bituminous Coal and Lignite.
}
\item{Anth.}{
a numeric vector, price of Anthracite.
}
}
}

\source{
 Darrel Sala, Institute of Gas Technology, Chicago.
}
\examples{
data(E7.2)
summary(E7.2)
}
\keyword{datasets}
\concept{regression}

