\name{E7.6}
\alias{E7.6}
\non_function{}
\title{ Community Area Data for the North Part of the City of Chicago }
\usage{data(E7.6)}
\description{
The \code{E7.6} data frame has 34 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Area.Name}{
a character vector, name of area.
}
\item{PB}{
a numeric vector, percentage of population which are black.
}
\item{PS}{
a numeric vector, percentage of population which are spanish speaking.
}
\item{PA}{
a numeric vector, percentage of population over 65.
}
\item{Income}{
a numeric vector, median family income for each area.
}
}
}
\seealso{ \code{\link{E7.7}}, which is the adjacency 
 matrix for the 34 areas. }
\source{
 Data set were constructed by Prof. Siim Soot, Dept. of Geography, 
 University of Illinois at Chicago. 
}
\examples{
data(E7.6)
}
\keyword{datasets}
