\name{E7.7}
\alias{E7.7}
\title{ The Contiguity Matrix for the 34 Areas in Northern Chicago  }
\concept{ The Contiguity Matrix for the 34 Areas in Northern Chicago  }
\usage{data(E7.7)}
\description{
 This is the contiguity matrix for the 34 areas in northern Chicago, 
 given in \code{\link{E7.6}}.  Contains only 0's and 1's with the 
 obvious interpretation.
}
\keyword{datasets}
