\name{E9.19}
\alias{E9.19}
\non_function{}
\title{ Acceleration data }
\usage{data(E9.19)}
\description{
The \code{E9.19} data frame has 50 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{ACC}{
a numeric vector, Acceleration of different vehicles.
}
\item{WHP}{
a numeric vector, weight-to-horsepower ratio.
}
\item{SP}{
a numeric vector, speed at which they were travelling.
}
\item{G}{
a numeric vector, Grade of road, G=0 implies road was horizontal.
}
}
}

\source{
 Raj Tejwaney, Department of civil Engineering, University
  of Illinoi at Chicago.
}
\examples{
data(E9.19)
}
\keyword{datasets}
