\name{E9.20}
\alias{E9.20}
\non_function{}
\title{ Stadium Cleanup Data }
\usage{data(E9.20)}
\description{
The \code{E9.20} data frame has 16 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C}{
a numeric vector, cost of cleanup. Units forgotten.
}
\item{R.HD}{
a numeric vector, sales at hot-dog stands. Units forgotten.
}
\item{R.B}{
a numeric vector, sales at beer stands. Units forgotten.
}
}
}
\source{
 The authors of the book.
}
\examples{
data(E9.20)
}
\keyword{datasets}
