\name{E9.21}
\alias{E9.21}
\title{ Depreciation in Market Value of Large Factories }
\concept{ Depreciation in Market Value of Large Factories }
\usage{data(E9.21)}
\description{
   The \code{E9.21} data frame has 11 rows and 2 columns.
}
\format{
   This data frame contains the following columns:
\describe{
\item{Age}{
a numeric vector, units not given, probably years.
}
\item{Depr}{
a numeric vector, units not given.
}
}
}
\source{
 Diamond-stars Motors, Normal, Il. Gary Shultz, general Counsel, made this 
       data available.
}
\examples{
data(E9.21)
summary(E9.21)
plot(E9.21)
}
\keyword{datasets}
\concept{regression}

